+++ Userland/Libraries/LibC/string.c
@@ -18,11 +18,11 @@ void* memsetw(void* ptr, int value, size_t count)
     return ptr;
 }
 
-void* memsetq(void* ptr, int value, size_t count)
+void* memsetq(void* ptr, uint64_t value, size_t count)
 {
     uint64_t* buf = (uint64_t*) ptr;
     for (size_t i = 0; i < count; ++i)
-        buf[i] = (uint64_t) value;
+        buf[i] = value;
 
     return ptr;
 }
